big push
This is a big push, i been working at home for the whole things so I never though to push until it ready. what done is the first four spell attacks, all enemies(not bosses) , level design, mana and health system. what next is going to be the art push for every thing, then ui in game